+++ ext/polars/src/functions/range.rs
@@ -39,3 +39,19 @@ pub fn date_range(
     let time_unit = time_unit.map(|x| x.0);
     dsl::date_range(start, end, every, closed, time_unit, time_zone).into()
 }
+
+pub fn date_ranges(
+    start: &RbExpr,
+    end: &RbExpr,
+    every: String,
+    closed: Wrap<ClosedWindow>,
+    time_unit: Option<Wrap<TimeUnit>>,
+    time_zone: Option<TimeZone>,
+) -> RbExpr {
+    let start = start.inner.clone();
+    let end = end.inner.clone();
+    let every = Duration::parse(&every);
+    let closed = closed.0;
+    let time_unit = time_unit.map(|x| x.0);
+    dsl::date_ranges(start, end, every, closed, time_unit, time_zone).into()
+}